services/fraud-detection-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
kafka-python-ng==2.2.2
py-eureka-client==0.11.3
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
requests==2.31.0
python-json-logger==2.0.7
pytest==7.4.3
httpx==0.25.2